+++ src/mcfacts/physics/binary/__init__.py
@@ -1,0 +1,3 @@
+"""
+Functions for handling binaries
+"""
