+++ src/mcfacts/physics/binary/evolve.py
@@ -1,9 +1,9 @@
+"""
+Functions for evolving the state of a binary
+"""
 import numpy as np
 import scipy
 from mcfacts.objects.agnobject import obj_to_binary_bh_array
-from astropy import constants as const
-from astropy import units as u
-from astropy.units import cds
 
 
 def change_bin_mass(blackholes_binary, disk_bh_eddington_ratio,
@@ -164,7 +164,7 @@ def change_bin_spin_angles(blackholes_binary, disk_bh_eddington_ratio,
     return (blackholes_binary)
 
 
-def com_feedback_hankla(blackholes_binary, disk_surface_density, disk_opacity_func, disk_bh_eddington_ratio, disk_alpha_viscosity, disk_radius_outer):
+def bin_com_feedback_hankla(blackholes_binary, disk_surface_density, disk_opacity_func, disk_bh_eddington_ratio, disk_alpha_viscosity, disk_radius_outer):
     """
     This feedback model uses Eqn. 28 in Hankla, Jiang & Armitage (2020)
     which yields the ratio of heating torque to migration torque.
@@ -389,207 +389,7 @@ def bin_migration_obj(smbh_mass, blackholes_binary, disk_surf_model, disk_aspect
     return (blackholes_binary)
 
 
-def gw_strain_freq(mass_1, mass_2, obj_sep, timestep_duration_yr, old_gw_freq, smbh_mass, agn_redshift, flag_include_old_gw_freq=1):
-    """
-    This function takes in two masses, their separation, the previous frequency, and the redshift and
-    calculates the new GW strain (unitless) and frequency (Hz).
-
-    Note from Saavik about hardcoding strain_factor to 4e3 if nu_gw > 1e-6:
-        basically we are implicitly assuming if the frequency is low enough the source is monochromatic
-        in LISA over the course of 1yr, so that's where those values come from... and we do need to make
-        a decision about that... and that's an ok decision for now. But if someone were to be considering
-        a different observatory they might not like that decision?
-
-    Parameters
-    ----------
-    mass_1 : numpy array
-        mass of object 1
-    mass_2 : numpy array
-        mass of object 2
-    obj_sep : numpy array
-        separation between both objects, in R_g of the SMBH = GM_smbh/c^2
-    timestep_duration_yr : float, or -1 if not given
-        current timestep in years
-    old_gw_freq : numpy array, or -1 if not given
-        previous GW frequency
-    smbh_mass : float
-        mass of the SMBH in Msun
-    agn_redshift : float
-        redshift of the SMBH
-    flag_include_old_gw_freq : boolean
-        flag indicating if old_gw_freq should be included in calculations
-        if not, we use the hardcoded value (see note above)
-        0 if no, 1 if yes
-
-    Returns
-    -------
-    char_strain : numpy array
-        dimensionless characteristic strain
-    nu_gw : numpy array
-        GW frequency in Hz
-    """
-
-    redshift_d_obs_dict = {0.1: 421*u.Mpc,
-                           0.5: 1909*u.Mpc}
-
-    timestep_units = (timestep_duration_yr*u.year).to(u.second)
-
-    # 1rg =1AU=1.5e11m for 1e8Msun
-    rg = 1.5e11*(smbh_mass/1.e8)*u.meter
-    mass_1 = (mass_1 * cds.Msun).to(u.kg)
-    mass_2 = (mass_2 * cds.Msun).to(u.kg)
-    mass_total = mass_1 + mass_2
-    bin_sep = obj_sep * rg
-
-    mass_chirp = np.power(mass_1 * mass_2, 3./5.) / np.power(mass_total, 1./5.)
-    rg_chirp = ((const.G * mass_chirp) / np.power(const.c, 2)).to(u.meter)
-
-    # If separation is less than rg_chirp then cap separation at rg_chirp.
-    bin_sep[bin_sep < rg_chirp] = rg_chirp[bin_sep < rg_chirp]
-
-    nu_gw = (1.0/np.pi) * np.sqrt(mass_total * const.G / np.power(bin_sep, 3))
-    nu_gw = nu_gw.to(u.Hz)
-
-    # For local distances, approx d=cz/H0 = 3e8m/s(z)/70km/s/Mpc =3.e8 (z)/7e4 Mpc =428 Mpc
-    # From Ned Wright's calculator https://www.astro.ucla.edu/~wright/CosmoCalc.html
-    # (z=0.1)=421Mpc. (z=0.5)=1909 Mpc
-    d_obs = redshift_d_obs_dict[agn_redshift].to(u.meter)
-    strain = (4/d_obs) * rg_chirp * np.power(np.pi * nu_gw * rg_chirp / const.c, 2./3.)
-    
-    # But power builds up in band over multiple cycles! 
-    # So characteristic strain amplitude measured by e.g. LISA is given by h_char^2 = N/8*h_0^2 where N is number of cycles per year & divide by 8 to average over viewing angles
-    strain_factor = np.ones(len(nu_gw))
-
-    # char amplitude = strain_factor*h0
-    #                = sqrt(N/8)*h_0 and N=freq*1yr for approx const. freq. sources over ~~yr.
-    strain_factor[nu_gw < (1e-6)*u.Hz] = np.sqrt(nu_gw[nu_gw < (1e-6)*u.Hz]*np.pi*(1e7)/8)
-
-    # For a source changing rapidly over 1 yr, N~freq^2/ (dfreq/dt).
-    # char amplitude = strain_factor*h0
-    #                = sqrt(freq^2/(dfreq/dt)/8)*h0
-    if (flag_include_old_gw_freq == 1):
-        delta_nu = np.abs(old_gw_freq - nu_gw)
-        delta_nu_delta_timestep = delta_nu/timestep_units
-        nu_squared = (nu_gw*nu_gw)
-        strain_factor[nu_gw > (1e-6)*u.Hz] = np.sqrt((nu_squared[nu_gw > (1e-6)*u.Hz] / delta_nu_delta_timestep[nu_gw > (1e-6)*u.Hz])/8.)
-    # Condition from evolve_gw
-    elif (flag_include_old_gw_freq == 0):
-        strain_factor[nu_gw > (1e-6)*u.Hz] = np.full(np.sum(nu_gw > (1e-6)*u.Hz), 4.e3)
-    char_strain = strain_factor*strain
-
-    return (char_strain.value, nu_gw.value)
-
-
-def evolve_gw(blackholes_binary, smbh_mass, agn_redshift):
-
-    char_strain, nu_gw = gw_strain_freq(mass_1=blackholes_binary.mass_1,
-                                        mass_2=blackholes_binary.mass_2,
-                                        obj_sep=blackholes_binary.bin_sep,
-                                        timestep_duration_yr=-1,
-                                        old_gw_freq=-1,
-                                        smbh_mass=smbh_mass,
-                                        agn_redshift=agn_redshift,
-                                        flag_include_old_gw_freq=0)
-
-    # Update binaries
-    blackholes_binary.gw_freq = nu_gw
-    blackholes_binary.gw_strain = char_strain
-
-    return (blackholes_binary)
-
-
-def bbh_gw_params(blackholes_binary, bh_binary_id_num_gw, smbh_mass, timestep_duration_yr, old_bbh_freq, agn_redshift):
-    """
-    This function evaluates the binary gravitational wave frequency and strain at the end of each timestep_duration_yr.
-    Set up binary GW frequency nu_gw = 1/pi *sqrt(GM_bin/a_bin^3). Set up binary strain of GW
-    h = (4/d_obs) *(GM_chirp/c^2)*(pi*nu_gw*GM_chirp/c^3)^(2/3)
-    where m_chirp =(M_1 M_2)^(3/5) /(M_bin)^(1/5)
-
-
-    Parameters
-    ----------
-    blackholes_binary : AGNBinaryBlackHole 
-        Full binary array.
-    bh_binary_id_num_gw : numpy array
-        ID numbers of binaries with separations below min_bbh_gw_separation
-    smbh_mass : float
-        Mass of the SMBH
-    timestep_duration_yr : float
-        timestep in years
-    old_bbh_freq : numpy array
-        Previous gw_freq
-    agn_redshift : float
-        Redshift of the AGN, used to set d_obs
-
-    Returns
-    -------
-    char_strain : numpy array
-        characteristic strain, unitless
-    nu_gw : numpy array
-        GW frequency in Hz
-    """
-
-    num_tracked = bh_binary_id_num_gw.size
-
-    old_bbh_freq = old_bbh_freq * u.Hz
-
-    while (num_tracked > len(old_bbh_freq)):
-        old_bbh_freq = np.append(old_bbh_freq, (9.e-7) * u.Hz)
-
-    while (num_tracked < len(old_bbh_freq)):
-        old_bbh_freq = np.delete(old_bbh_freq, 0)
-
-    char_strain, nu_gw = gw_strain_freq(mass_1=blackholes_binary.at_id_num(bh_binary_id_num_gw, "mass_1"),
-                                        mass_2=blackholes_binary.at_id_num(bh_binary_id_num_gw, "mass_2"),
-                                        obj_sep=blackholes_binary.at_id_num(bh_binary_id_num_gw, "bin_sep"),
-                                        timestep_duration_yr=timestep_duration_yr,
-                                        old_gw_freq=old_bbh_freq,
-                                        smbh_mass=smbh_mass,
-                                        agn_redshift=agn_redshift,
-                                        flag_include_old_gw_freq=1)
-
-    return (char_strain, nu_gw)
-
-
-def evolve_emri_gw(blackholes_inner_disk, timestep_duration_yr, old_gw_freq, smbh_mass, agn_redshift):
-    """
-    This function evaluates the EMRI gravitational wave frequency and strain at the end of each timestep_duration_yr
-
-    Parameters
-    ----------
-    blackholes_inner_disk : AGNBlackHole
-        black holes in the inner disk
-    timestep_duration_yr : float
-        timestep in years
-    old_gw_freq : numpy array
-        previous GW frequency in Hz
-    smbh_mass : float
-        mass of the SMBH in Msun
-    agn_redshift : float
-        redshift of the AGN
-    """
-
-    old_gw_freq = old_gw_freq * u.Hz
-
-    # If number of EMRIs has grown since last timestep_duration_yr, add a new component to old_gw_freq to carry out dnu/dt calculation
-    while (blackholes_inner_disk.num < len(old_gw_freq)):
-        old_gw_freq = np.delete(old_gw_freq, 0)
-    while blackholes_inner_disk.num > len(old_gw_freq):
-        old_gw_freq = np.append(old_gw_freq, (9.e-7) * u.Hz)
-
-    char_strain, nu_gw = gw_strain_freq(mass_1=smbh_mass,
-                                        mass_2=blackholes_inner_disk.mass,
-                                        obj_sep=blackholes_inner_disk.orb_a,
-                                        timestep_duration_yr=timestep_duration_yr,
-                                        old_gw_freq=old_gw_freq,
-                                        smbh_mass=smbh_mass,
-                                        agn_redshift=agn_redshift,
-                                        flag_include_old_gw_freq=1)
-
-    return (char_strain, nu_gw)
-
-
-def ionization_check(blackholes_binary, smbh_mass):
+def bin_ionization_check(blackholes_binary, smbh_mass):
     """
     This function tests whether a binary has been softened beyond some limit.
     Returns ID numbers of binaries to be ionized.
@@ -637,7 +437,7 @@ def ionization_check(blackholes_binary, smbh_mass):
     return (bh_id_nums)
 
 
-def contact_check(blackholes_binary, smbh_mass):
+def bin_contact_check(blackholes_binary, smbh_mass):
     """
     This function tests to see if the binary separation has shrunk so that the binary is touching!
 
@@ -674,7 +474,7 @@ def contact_check(blackholes_binary, smbh_mass):
     return (blackholes_binary)
 
 
-def reality_check(blackholes_binary):
+def bin_reality_check(blackholes_binary):
     """ This function tests to see if the binary is real. If location = 0 or mass = 0 *and* any other element is NON-ZERO then discard this binary element.
         Returns ID numbers of fake binaries.
 
@@ -702,3 +502,97 @@ def reality_check(blackholes_binary):
         return (id_nums)
     else:
         return (bh_bin_id_num_fakes)
+
+
+def bin_harden_baruteau(blackholes_binary, smbh_mass, timestep_duration_yr,
+                        time_gw_normalization, time_passed):
+    """
+    Harden black hole binaries using Baruteau+11 prescription
+
+    Use Baruteau+11 prescription to harden a pre-existing binary.
+    For every 1000 orbits of binary around its center of mass, the
+    separation (between binary components) is halved.
+
+    Parameters
+    ----------
+    binary_bh_array : ndarray
+        Array of binary black holes in the disk.
+    smbh_mass : ndarray
+        Mass of supermassive black hole.
+    timestep_duration_yr : float
+        Length of timestep of the simulation in years.
+    time_gw_normalization : float
+        A normalization for GW decay timescale, set by `smbh_mass` & normalized for
+        a binary total mass of 10 solar masses.
+    bin_index : int
+        Count of number of binaries
+    time_passed : float
+        Time elapsed since beginning of simulation.
+
+    Returns
+    -------
+    ndarray
+        Updated array of black hole binaries in the disk.
+    """
+
+    # 1. Find active binaries
+    # 2. Find number of binary orbits around its center of mass within the timestep
+    # 3. For every 10^3 orbits, halve the binary separation.
+
+
+    # Only interested in BH that have not merged
+    idx_non_mergers = np.where(blackholes_binary.flag_merging >= 0)[0]
+
+    # If all binaries have merged then nothing to do
+    if (idx_non_mergers.shape[0] == 0):
+        return blackholes_binary
+
+    # Set up variables
+    mass_binary = blackholes_binary.mass_1[idx_non_mergers] + blackholes_binary.mass_2[idx_non_mergers]
+    mass_reduced = (blackholes_binary.mass_1[idx_non_mergers] * blackholes_binary.mass_2[idx_non_mergers]) / mass_binary
+    bin_sep = blackholes_binary.bin_sep[idx_non_mergers]
+    bin_orb_ecc = blackholes_binary.bin_ecc[idx_non_mergers]
+
+    # Find eccentricity factor (1-e_b^2)^7/2
+    ecc_factor_1 = np.power(1 - np.power(bin_orb_ecc, 2), 3.5)
+    # and eccentricity factor [1+(73/24)e_b^2+(37/96)e_b^4]
+    ecc_factor_2 = 1 + ((73/24) * np.power(bin_orb_ecc, 2)) + ((37/96) * np.power(bin_orb_ecc, 4))
+    # overall ecc factor = ecc_factor_1/ecc_factor_2
+    ecc_factor = ecc_factor_1/ecc_factor_2
+
+    # Binary period = 2pi*sqrt((delta_r)^3/GM_bin)
+    # or T_orb = 10^7s*(1r_g/m_smmbh=10^8Msun)^(3/2) *(M_bin/10Msun)^(-1/2) = 0.32yrs
+    bin_period = 0.32 * np.power(bin_sep, 1.5) * np.power(smbh_mass/1.e8, 1.5) * np.power(mass_binary/10.0, -0.5)
+
+    # Find how many binary orbits in timestep. Binary separation is halved for every 10^3 orbits.
+    num_orbits_in_timestep = np.zeros(len(bin_period))
+    num_orbits_in_timestep[bin_period > 0] = timestep_duration_yr / bin_period[bin_period > 0]
+    scaled_num_orbits = num_orbits_in_timestep / 1000.0
+
+    # Timescale for binary merger via GW emission alone, scaled to bin parameters
+    time_to_merger_gw = time_gw_normalization*((bin_sep)**(4.0))*((mass_binary/10.0)**(-2))*((mass_reduced/2.5)**(-1.0))*ecc_factor
+    # Finite check
+    assert np.isfinite(time_to_merger_gw).all(),\
+        "Finite check failure: time_to_merger_gw"
+    blackholes_binary.time_to_merger_gw[idx_non_mergers] = time_to_merger_gw
+
+    # Binary will not merge in this timestep
+    # new bin_sep according to Baruteu+11 prescription
+    bin_sep[time_to_merger_gw > timestep_duration_yr] = bin_sep[time_to_merger_gw > timestep_duration_yr] * np.power(0.5, scaled_num_orbits[time_to_merger_gw > timestep_duration_yr])
+    blackholes_binary.bin_sep[idx_non_mergers[time_to_merger_gw > timestep_duration_yr]] = bin_sep[time_to_merger_gw > timestep_duration_yr]
+    # Finite check
+    assert np.isfinite(blackholes_binary.bin_sep).all(),\
+        "Finite check failure: blackholes_binary.bin_sep"
+
+    # Otherwise binary will merge in this timestep
+    # Update flag_merging to -2 and time_merged to current time
+    blackholes_binary.flag_merging[idx_non_mergers[time_to_merger_gw <= timestep_duration_yr]] = np.full(np.sum(time_to_merger_gw <= timestep_duration_yr), -2)
+    blackholes_binary.time_merged[idx_non_mergers[time_to_merger_gw <= timestep_duration_yr]] = np.full(np.sum(time_to_merger_gw <= timestep_duration_yr), time_passed)
+    # Finite check
+    assert np.isfinite(blackholes_binary.flag_merging).all(),\
+        "Finite check failure: blackholes_binary.flag_merging"
+    # Finite check
+    assert np.isfinite(blackholes_binary.time_merged).all(),\
+        "Finite check failure: blackholes_binary.time_merged"
+
+    return (blackholes_binary)
