+++ src/mcfacts/physics/binary/merge.py
@@ -1,4 +1,10 @@
+"""
+Functions for calculating final variables of a merging binary
+"""
 import numpy as np
+from astropy import units as u
+
+from mcfacts.physics.point_masses import time_of_orbital_shrinkage, si_from_r_g
 
 
 def chi_effective(masses_1, masses_2, spins_1, spins_2, spin_angles_1, spin_angles_2, bin_ang_mom):
@@ -8,7 +14,7 @@ def chi_effective(masses_1, masses_2, spins_1, spins_2, spin_angles_1, spin_angl
     The measured effective spin of a merger is calculated as
 
     .. math:: \chi_{\rm eff}=\frac{m_1*\chi_1*\cos(\theta_1) + m_2*\chi_2*\cos(\theta_2)}/{m_{\rm bin}}
-    
+
     Parameters
     ----------
     masses_1 : numpy array
@@ -123,3 +129,135 @@ def chi_p(masses_1, masses_2, spins_1, spins_2, spin_angles_1, spin_angles_2, bi
     chi_p[chi_p < (mass_ratio_factors * spins_2_perp)] = mass_ratio_factors[chi_p < (mass_ratio_factors * spins_2_perp)] * spins_2_perp[chi_p < (mass_ratio_factors * spins_2_perp)]
 
     return (chi_p)
+
+
+def normalize_tgw(smbh_mass, inner_disk_outer_radius):
+    """Gravitational wave timescale normalization.
+
+    Calculate the normalization for timescale of a merger (in yrs) due to GW emission.
+    From Peters(1964):
+
+    t_gw approx (5/256)* c^5/G^3 *a_b^4/(M_{b}^{2}mu_{b}) assuming ecc=0.0.
+
+    For a_b in units of r_g=GM_smbh/c^2 we find
+
+    t_gw=(5/256)*(G/c^3)*(a/r_g)^{4} *(M_s^4)/(M_b^{2}mu_b)
+
+    Put bin_mass_ref in units of 10Msun is a reference mass.
+    reduced_mass in units of 2.5Msun.
+
+    Parameters
+    ----------
+    smbh_mass : float
+        Mass of the supermassive black hole.
+    inner_disk_outer_radius : float
+        Outer radius of the inner disk (r_g)
+
+    Returns
+    -------
+    float
+        normalization to gravitational wave timescale
+    """
+
+    bin_mass_ref = 10.0
+    '''
+    G = const.G
+    c = const.c
+    mass_sun = const.M_sun
+    year = 3.1536e7
+    reduced_mass = 2.5
+    norm = (5.0/256.0)*(G/(c**(3)))*(smbh_mass**(4))*mass_sun/((bin_mass_ref**(2))*reduced_mass)
+    time_gw_normalization = norm/year
+    '''
+    time_gw_normalization = time_of_orbital_shrinkage(
+        smbh_mass * u.solMass,
+        bin_mass_ref * u.solMass,
+        si_from_r_g(smbh_mass * u.solMass, inner_disk_outer_radius),
+        0 * u.m,
+    )
+    return time_gw_normalization.si.value
+
+
+def merged_mass(masses_1, masses_2, spins_1, spins_2):
+    """
+    Calculate the final mass of a merged binary.
+
+    Using approximations from Tichy \& Maronetti (2008) where
+    m_final=(M1+M2)*[1.0-0.2nu-0.208nu^2(a1+a2)]
+    where nu is the symmetric mass ratio or nu=q/((1+q)^2)
+
+    Parameters
+    ----------
+    masses_1 : numpy array
+        Masses of objects 1.
+    masses_2 : numpy array
+        Masses of objects 2.
+    spins_1 : numpy array
+        Spin magnitudes of objects 1.
+    spins_2 : numpy array
+        Spin magnitudes of objects 2.
+
+    Returns
+    -------
+    merged_masses: numpy array
+        Final mass of merger remnant.
+    """
+
+    mass_ratios = np.ones(masses_1.size)
+    mass_ratios[masses_1 > masses_2] = masses_2[masses_1 > masses_2] / masses_1[masses_1 > masses_2]
+    mass_ratios[masses_1 < masses_2] = masses_1[masses_1 < masses_2] / masses_2[masses_1 < masses_2]
+
+    total_masses = masses_1 + masses_2
+    total_spins = spins_1 + spins_2
+    nu_factors = np.power(1.0 + mass_ratios, 2)
+    nu = mass_ratios / nu_factors
+    nu_squared = nu * nu
+
+    mass_factors = 1.0 - (0.2 * nu) - (0.208 * nu_squared * total_spins)
+    merged_masses = total_masses*mass_factors
+
+    return (merged_masses)
+
+
+def merged_spin(masses_1, masses_2, spins_1, spins_2):
+    """
+    Calculate the spin magnitude of a merged binary.
+
+    Only depends on M1,M2,a1,a2 and the binary ang mom around its center of mass.
+    Using approximations from Tichy \& Maronetti(2008) where
+    a_final=0.686(5.04nu-4.16nu^2) +0.4[a1/((0.632+1/q)^2)+ a2/((0.632+q)^2)]
+    where q=m_2/m_1 and nu=q/((1+q)^2)
+
+    Parameters
+    ----------
+    masses_1 : numpy array
+        Masses of objects 1.
+    masses_2 : numpy array
+        Masses of objects 2.
+    spins_1 : numpy array
+        Spin magnitudes of objects 1.
+    spins_2 : numpy array
+        Spin magnitudes of objects 2.
+
+    Returns
+    -------
+    merged_spins : numpy array
+        Final spin magnitude of merger remnant.
+    """
+
+    mass_ratios = np.ones(masses_1.size)
+    mass_ratios[masses_1 > masses_2] = masses_2[masses_1 > masses_2] / masses_1[masses_1 > masses_2]
+    mass_ratios[masses_1 < masses_2] = masses_1[masses_1 < masses_2] / masses_2[masses_1 < masses_2]
+
+    mass_ratios_inv = 1.0 / mass_ratios
+
+    nu_factors = np.power(1.0 + mass_ratios, 2)
+    nu = mass_ratios / nu_factors
+    nu_squared = nu * nu
+
+    spin_factors_1 = np.power(0.632 + mass_ratios_inv, 2)
+    spin_factors_2 = np.power(0.632 + mass_ratios, 2)
+
+    merged_spins = 0.686 * ((5.04 * nu) - (4.16 * nu_squared)) + (0.4 * ((spins_1 / spin_factors_1) + (spins_2 / spin_factors_2)))
+
+    return (merged_spins)
