+++ src/mcfacts/physics/point_masses.py
